pyproject.toml
[project]
name = "fixer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yoandy Rodriguez Martinez", email = "yoandy.rodriguez@musala.com" }
]
dependencies = [
    "python-frontmatter>=1.1.0",
    "mistletoe>=1.3.0",
    "click>=8.1.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "fixer:fix"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "python-lsp-server[all]>=1.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fixer"]
